+++ product/module.go
@@ -48,8 +48,8 @@ func (m *Module) Configure(injector *dingo.Injector) {
 
 	injector.BindMulti(new(graphql.Service)).To(new(productgraphql.Service))
 	if m.fakeService {
-		injector.Bind((*domain.ProductService)(nil)).To(fake.ProductService{})
-		injector.Bind((*domain.SearchService)(nil)).To(fake.SearchService{})
+		injector.Bind((*domain.ProductService)(nil)).To(fake.ProductService{}).In(dingo.ChildSingleton)
+		injector.Bind((*domain.SearchService)(nil)).To(fake.SearchService{}).In(dingo.ChildSingleton)
 	}
 
 }
@@ -75,7 +75,10 @@ commerce: {
 		slugAttributeCode: string | *"urlSlug"
 		fakeservice: {
 			enabled: bool | *false
-			currency: *"€" | !=""  
+			currency: *"€" | !=""
+			if enabled {
+			  jsonTestDataFolder: string | *"testdata/products"
+			}
 		}
 		api: {
 			enabled: bool | *true